Refactor of Solid Helpers and minor gui changes
