+++ webapp/src/localize/en/translation.json
@@ -91,7 +91,6 @@
         "show": "Show friends",
         "addB": "Add",
         "cancel": "Cancel",
-        "noLoggedIn": "Log-In to see your friends",
-        "loading": "Loading friends..."
+        "noLoggedIn": "Log-In to see your friends"
     }
 }
