+++ webapp/src/localize/es/translation.json
@@ -91,7 +91,6 @@
         "show": "Ver amigos",
         "addB": "Añadir",
         "cancel": "Cancelar",
-        "noLoggedIn": "Inicia sesión para ver a tus amigos",
-        "loading": "Cargando amigos..."
+        "noLoggedIn": "Inicia sesión para ver a tus amigos"
     }
 }
